Thank you very much!

Actually, the gcc version warning is because I am using `gcc 5.3.1`.

Respect the other fixes, nothing to say, just thanks for taking your time do fix them :) :+1:

+++ utils/utils.c
@@ -152,8 +152,8 @@ int readDenseCoordinateMatrixRowLine(char *fileName,unsigned long **I,unsigned l
 	*J = (unsigned long *)  malloc(*nz * sizeof(unsigned long));
 	*values = (double *)  malloc(*nz * sizeof(double));
 	
-	unsigned int *tmpI = *I;
-	unsigned int *tmpJ = *J;
+	unsigned long *tmpI = *I;
+	unsigned long *tmpJ = *J;
 	double *tmpValues = *values;
 	
 	//Read rows
